refactor: separate sasl and scram concept
